requirements.txt
torch~=1.8.1
chainer~=1.24.0
scikit-image~=0.18.1
numpy~=1.20.2
matplotlib~=3.4.1
argparse~=1.4.0
Pillow~=8.2.0
IPython~=7.22.0
PySimpleGUI~=4.40.0